refactor(uzicli): split runs-domain verbs into client_runs.go/fake_runs.go (PRD #1017 M1)
Pure code motion: move the run-lifecycle HTTPClient verbs and their FakeClient
twins (plus fake-side StreamRun) out of client.go/fake.go into per-domain files.
go doc -u -all ./internal/uzicli byte-identical; no test edits.
